nothing added, issue found
all the enemies are on the same timer not their individual ones.
this is happenin cause they're all on the same script and that script is just looking for a timer rather than the one attached to the actual enemy.
not entirely sure what the best way to fix it is but theres prob a way to just look for a timer attached to the node